Lifecycle Hooks and Instance Status
Previously, we had a "dirty" property in Component Instance, which was used to determine wether a Component should be re-rendered or not, this has been replaced with a Instance Status with the 3 following options:
BEFORE_MOUNT (0) - The initial state of all Component Instances, signifies that the Component is out of sync (aka. dirty) and is newly mounted (so therefore call the onMounted hook)
SYNCED (1) - The component is in sync, will not render unless render is ordered by the Internal Render Engine through patchComponent (if the props has changed).
UNSYNCED (2) - The component is out of sync, re-render when possible.

Also added the following lifecycle hooks:
onCreated (formerly instanceSetup) - Called when Component Instance is created, before any Render Function is called (useful for setting up values used in Render Functions)
onMounted - Called on first render (when status is BEFORE_MOUNT)
onUpdated - Called on other renders (when status is not BEFORE_MOUNT)
onDestroy - Called after Component has been unmounted, and before Component Instance has been destroyed.